scripts/requirements.txt
Pillow==3.4.1
pyaml==16.9.0
PyYAML==3.12
Wand==0.4.3
xlrd==1.0.0
